Adding disclaimer to the login page and moved into own component
