+++ frontend/src/pages/mod.rs
@@ -2,8 +2,10 @@
 //! To export pages
 mod dashboard;
 mod home;
+mod login;
 mod new_user;
 
 pub use dashboard::*;
 pub use home::*;
+pub use login::*;
 pub use new_user::*;
